chore: add Drupal AI Initiative partners demo hosted form
