+++ app/Forms/DrupalAIDemoDrupalOrgForm.php
@@ -2,9 +2,6 @@
 
 namespace App\Forms;
 
-use App\Enums\PolydockStoreAppStatusEnum;
-use App\Enums\PolydockStoreStatusEnum;
-use App\Rules\BannedEmail;
 use Illuminate\Validation\Rule;
 
 class DrupalAIDemoDrupalOrgForm extends BaseHostedForm
@@ -52,10 +49,7 @@ class DrupalAIDemoDrupalOrgForm extends BaseHostedForm
     #[\Override]
     public function getValidationRules(): array
     {
-        return [
-            'first_name' => ['required', 'string', 'max:100'],
-            'last_name' => ['required', 'string', 'max:100'],
-            'email' => ['required', 'email', new BannedEmail],
+        return array_merge(parent::getValidationRules(), [
             'organization' => ['nullable', 'string', 'max:150'],
             'job_title' => ['nullable', 'string', 'max:150'],
             'country' => [
@@ -65,22 +59,7 @@ class DrupalAIDemoDrupalOrgForm extends BaseHostedForm
             ],
             'stage_in_ai_adoption' => ['nullable', 'string', 'in:just-curious,specific-need,already-using'],
             'interest_in_drupal_ai' => ['nullable', 'string', 'max:255'],
-            'trial_app' => [
-                'required',
-                'uuid',
-                Rule::exists('polydock_store_apps', 'uuid')
-                    ->where('status', PolydockStoreAppStatusEnum::AVAILABLE->value)
-                    ->where('available_for_trials', true)
-                    ->where(function ($query) {
-                        $query->whereExists(function ($subQuery) {
-                            $subQuery->selectRaw(1)
-                                ->from('polydock_stores')
-                                ->whereColumn('polydock_stores.id', 'polydock_store_apps.polydock_store_id')
-                                ->where('polydock_stores.status', PolydockStoreStatusEnum::PUBLIC->value);
-                        });
-                    }),
-            ],
-        ];
+        ]);
     }
 
     #[\Override]
