+++ resources/views/forms/partials/contact-fields.blade.php
@@ -1,0 +1,20 @@
+<div class="form-group">
+    <label class="form-label" for="firstName">
+        First Name <span class="required-star">*</span>
+    </label>
+    <input type="text" id="firstName" name="first_name" required class="form-control" autocomplete="given-name">
+</div>
+
+<div class="form-group">
+    <label class="form-label" for="lastName">
+        Last Name <span class="required-star">*</span>
+    </label>
+    <input type="text" id="lastName" name="last_name" required class="form-control" autocomplete="family-name">
+</div>
+
+<div class="form-group">
+    <label class="form-label" for="email">
+        Work Email <span class="required-star">*</span>
+    </label>
+    <input type="email" id="email" name="email" required class="form-control" autocomplete="email">
+</div>
